build: Publish package in publish job in ci.yml (#31)
In the publish job in the ci.yml workflow, we create a new release using
the GitHub CLI. As we are using the GITHUB_TOKEN, this will not trigger
the release.yml workflow. I copied the logic to ci.yml. In a real-world
application, it would be a composite action or reusable workflow.